refactor: extract MCP document mutation handlers
